+++ packages/message-lint-rules/emptyPattern/src/emptyPattern.test.ts
@@ -6,6 +6,7 @@ import { lintSingleMessage } from "@inlang/sdk/lint"
 
 const message1: Message = {
 	id: "1",
+	alias: {},
 	selectors: [],
 	variants: [
 		{ languageTag: "en", match: [], pattern: [{ type: "Text", value: "Inlang" }] },