+++ packages/sdk/src/adapter/solidAdapter.test.ts
@@ -28,6 +28,11 @@ const config: ProjectSettings = {
 	},
 }
 
+const configWithAliases: ProjectSettings = {
+	...config,
+	experimental: { aliases: true },
+}
+
 const mockPlugin: Plugin = {
 	id: "plugin.project.i18next",
 	description: { en: "Mock plugin description" },
@@ -41,6 +46,7 @@ const mockPlugin: Plugin = {
 const exampleMessages: Message[] = [
 	{
 		id: "a",
+		alias: {},
 		selectors: [],
 		variants: [
 			{
@@ -57,6 +63,7 @@ const exampleMessages: Message[] = [
 	},
 	{
 		id: "b",
+		alias: {},
 		selectors: [],
 		variants: [
 			{
@@ -202,10 +209,10 @@ describe("messages", () => {
 			{ from }
 		)
 
-		let counter = 0
+		let effectOnMessagesCounter = 0
 		createEffect(() => {
 			project.query.messages.getAll()
-			counter += 1
+			effectOnMessagesCounter += 1
 		})
 
 		expect(Object.values(project.query.messages.getAll()).length).toBe(2)
@@ -215,11 +222,11 @@ describe("messages", () => {
 		// TODO: how can we await `setConfig` correctly
 		await new Promise((resolve) => setTimeout(resolve, 510))
 
-		expect(counter).toBe(1) // 2 times because effect creation + set
+		expect(effectOnMessagesCounter).toBe(1) // 2 times because effect creation + set
 		expect(Object.values(project.query.messages.getAll()).length).toBe(2)
 	})
 
-	it("should react to changes in messages", async () => {
+	it("should react to message udpate", async () => {
 		const repo = await mockRepo()
 		const fs = repo.nodeishFs
 		await fs.mkdir("/user/project.inlang.inlang", { recursive: true })
@@ -266,6 +273,71 @@ describe("messages", () => {
 					},
 				],
 			},
+		})
+
+		it("should react to message udpate (with aliases)", async () => {
+			const repo = await mockRepo()
+			const fs = repo.nodeishFs
+			await fs.mkdir("/user/project.inlang.inlang", { recursive: true })
+			await fs.writeFile(
+				"/user/project.inlang.inlang/settings.json",
+				JSON.stringify(configWithAliases)
+			)
+			const project = solidAdapter(
+				await loadProject({
+					projectPath: "/user/project.inlang.inlang",
+					repo,
+					_import: $import,
+				}),
+				{ from }
+			)
+
+			let counter = 0
+			createEffect(() => {
+				project.query.messages.getAll()
+				counter += 1
+			})
+
+			const messagesBefore = project.query.messages.getAll
+			expect(Object.values(messagesBefore()).length).toBe(2)
+			expect(
+				(
+					Object.values(messagesBefore())[0]?.variants.find(
+						(variant) => variant.languageTag === "en"
+					)?.pattern[0] as Text
+				).value
+			).toBe("test")
+
+			project.query.messages.update({
+				where: { id: "raw_tapir_pause_grateful" },
+				// TODO: use `createMessage` utility
+				data: {
+					...exampleMessages[0],
+					variants: [
+						{
+							languageTag: "en",
+							match: [],
+							pattern: [
+								{
+									type: "Text",
+									value: "test2",
+								},
+							],
+						},
+					],
+				},
+			})
+
+			expect(counter).toBe(2) // 2 times because effect creation + set
+			const messagesAfter = project.query.messages.getAll
+			expect(Object.values(messagesAfter()).length).toBe(2)
+			expect(
+				(
+					Object.values(messagesAfter())[0]?.variants.find(
+						(variant) => variant.languageTag === "en"
+					)?.pattern[0] as Text
+				).value
+			).toBe("test2")
 		})
 
 		expect(counter).toBe(2) // 2 times because effect creation + set